initial

+++ cifar100.py
@@ -91,7 +91,7 @@ class ResNet50Classifier(pl.LightningModule):
 def main():
     data_module = CIFAR100DataModule(batch_size=64)
     model = ResNet50Classifier()
-    trainer = Trainer(accelerator='npu', devices='0,1', max_epochs=5, strategy='deepspeed', precision=16)
+    trainer = Trainer(accelerator='npu', devices='0,1', max_epochs=5, strategy='deepspeed')
     trainer.fit(model, datamodule=data_module)
     trainer.test(model, datamodule=data_module)
